==> packages/pimms-learn/pimms_learn-0.4.2.tar.gz/pimms_learn-0.4.2/project/01_0_transform_data_to_wide_format.ipynb <==
# Transfer data for NAGuideR format


import pandas as pd

import vaep
import vaep.models
from vaep.io import datasplits

# catch passed parameters
args = None
args = dict(globals()).keys()

Papermill script parameters:

# files and folders
# Datasplit folder with data for experiment
folder_experiment: str = 'runs/example'
folder_data: str = ''  # specify data directory if needed
file_format_in: str = 'csv'  # file format of original splits, default pickle (pkl)
file_format_out: str = 'csv'  # file format of transformed splits, default csv

args = vaep.nb.get_params(args, globals=globals())
args

params = vaep.nb.args_from_dict(args)
# params = OmegaConf.create(args)
params

splits = datasplits.DataSplits.from_folder(params.data, file_format=params.file_format_in)

train_data = splits.train_X.unstack()
train_data

Save placeholder sample annotation for use in NAGuideR app which requires such a file

annotation = pd.Series('test', train_data.index).to_frame('group')
annotation.index.name = 'Samples'
annotation

fname = params.data / 'sample_annotation_placeholder.csv'
annotation.to_csv(fname)
fname

# Save with samples in columns

fname = params.data / 'data_wide_sample_cols.csv'
# fillna('Filtered')
train_data.T.to_csv(fname)
fname

# 'data_wide_sample_cols.csv'
